feat: move eth confirmation next to eth client so it can be reused
